+++ CHANGELOG.md
@@ -19,3 +19,11 @@
 ### Fixed
 
 * Update condition for auto validation
+
+## 0.0.4
+
+### Breaking changes
+
+Replace `TFDropdownField` property: `items`(`List<String>`) 👉 `options`(`Map<String, String>`)
+Rename `TFDropdownField` property: `initialItem` 👉 `initialValue`)
+